"Update AuthConfig to fix redirect issues
The GET request matcher is updated to include a wildcard character for "/items" so that all pages under this directory will be covered.
The defaultSuccessUrl for both formLogin and oauth2Login is altered to use a flag to always redirect to "/success" and "/oauth2-success" respectively after successful login. This rectifies a previously observed problem where users were not correctly redirected post-authentication."